requirements.txt
requests>=2.18.4
google-api-python-client>=1.7.7
google-auth-httplib2>=0.0.3
google-auth-oauthlib>=0.2.0
PyDrive>=1.3.1
urlextract>=0.8.3
python-dotenv>=0.10.1
vk_api>=11.3.0
python-telegram-bot>=0.0.1
